refactor(AppLayout): move PageWrapper logic to _app. simplify AuthenticatedPage
