+++ src/components/AppLayout/index.js
@@ -1,3 +1,2 @@
 export { default as withStripe } from './withStripe'
-export { default as asPage } from './Page'
-export { default as asAuthenticatedPage } from './AuthenticatedPage'
+export { default as authenticated } from './AuthenticatedPage'
